// src/app/model/product/enums/filter-set-modal-status.enum.ts
export const enum FilterSetModalStatus {
    DeletedSuccessfully = 'DeletedSuccessfully',
    DeletionFailed = 'DeletionFailed',
    AddedSuccessfully = 'AddedSuccessfully',
    AddingFailed = 'AddingFailed',
    NameUpdatedSuccessfully = 'NameUpdatedSuccessfully',
    UpdatingNameFailed = 'UpdatingNameFailed',
}